requirements.txt
Python libraries required for this project to run.
Install them using: pip install -r requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1